+++ host/src/commonMain/kotlin/include/sys/StructStatExt.kt
@@ -27,18 +27,18 @@ public fun StructStat.packTo(sink: Sink): Unit = sink.run {
     writeIntLe(blocks.toInt()) // 36
 
     accessTime.timeMillis.let {
-        writeLongLe((it / 1000U).toLong()) // 40
-        writeIntLe((1000U * (it % 1000U)).toInt()) // 48
+        writeLongLe((it / 1000)) // 40
+        writeIntLe((1000 * (it % 1000)).toInt()) // 48
         writeIntLe(0) // 52, padding
     }
     modificationTime.timeMillis.let {
-        writeLongLe((it / 1000U).toLong()) // 56
-        writeIntLe((1000U * (it % 1000U)).toInt()) // 64
+        writeLongLe((it / 1000)) // 56
+        writeIntLe((1000 * (it % 1000)).toInt()) // 64
         writeIntLe(0) // 68, padding
     }
     changeStatusTime.timeMillis.let {
-        writeLongLe((it / 1000U).toLong()) // 72
-        writeIntLe((1000U * (it % 1000U)).toInt()) // 80
+        writeLongLe((it / 1000)) // 72
+        writeIntLe((1000 * (it % 1000)).toInt()) // 80
         writeIntLe(0) // 84, padding
     }
     writeLongLe(inode.toLong()) // 88
